layout: Add check about whether session should start on Overview
In situations that there are UI applications that are meant to be
started together with the session, it is undesirable to boot into
the overview.

In ideal situations, we would have session management infrastructure
in place that would allow GNOME Shell to find out there's UI services
to be brought up in the startup phase, and be able to decide window
positions and workspaces beforehand.

But we don't have any of that. Add a sneaky file existence test
so the behavior can be tweaked by savvy users and e.g. installers.
The hope being that this is a temporary solution till we have the
ideal framework in place.

Related: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/3854